+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Version 0.4
 Currently in development.
 
 - Added posting of external links for objects
+- Added schema editor
 
 
 Version 0.3